Feat: Tidy files & modularise the other scripts
Removed Files
- dbRefactor.html (refactor complete)
- DBHelperMod.js (renamed to DBHelper)
- FilterModel.js (now a module inside /modules)
- ItemCreateHelper.js (now a module inside /modules)
- refactor.js (refactor complete)

Index.html
- remove script tag getting old ItemCreateHelper script

DBHelper.js
- Renamed the modified version to replace the old one

app.js
- importing ItemCreateHelper module

ItemCreateHelper.js
- now an ES6 module in the /modules folder

ServerDataHelper
- update import of dataModel
