--- src/html/dbRefactor.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en-gb">
-  <head>
-    <script type="application/javascript" charset="utf-8" src="scripts/idb.js"></script>
-    <script type="module">
-      import recorderApp from './../scripts/refactor.js';
-      document.recorderApp = recorderApp;
-    </script>
-  </head>
-
-  <body>
-    <h1> Refactoring this thing</h1>
-  </body> 
-
-</html>
