+++ src/html/index.html
@@ -88,7 +88,6 @@
     <script type="application/javascript" charset="utf-8" src="scripts/idb.js"></script>
     <script type="application/javascript" charset="utf-8" src="scripts/StudentSelectModel.js"></script>
     <script type="application/javascript" charset="utf-8" src="scripts/ServiceWorkerHelper.js"></script>
-    <script type="application/javascript" charset="utf-8" src="scripts/ItemCreateHelper.js"></script>
     <script type="module">
       import app from "/scripts/app.js";
       document.myApp = app
