--- src/scripts/refactor.js
@@ -1,9 +1,0 @@
-'use strict'  
-
-import DbHelperMod from './DbHelperMod.js';
-
-let dbHelper = new DbHelperMod;
-
-export default {
-  dbHelper
-}
